refactor backup_data
